fix(stt): collapse repetition loops at the final stop-transcript boundary
Launch-blocking saved-transcript integrity fix for the live Private sample path.
TranscriptionService.stopTranscription() selects the longer of the strategy
transcript vs the streaming-accumulated visibleTranscript; a doubled streaming
transcript can win by length and bypass PrivateWhisper's per-segment collapse,
persisting a near_whole_doubling transcript as service_result/selectedForSave
(observed live: "We should literally like, wait, um, basically" x2).

- Move the conservative collapseTranscriptRepetitionLoops guard from PrivateWhisper
  to @/utils/repetitionRisk (pure util; also avoids a circular/heavy service->mode import).
- Apply it to the final selected transcript in stopTranscription so the persisted
  value can never contain a whole-text/multi-word loop, regardless of source.
- Reproduce-first regression in TranscriptionService.pause.test.ts (verified fails
  without the guard).

Not a #770 regression -- pre-existing v4/Private assembly behavior, first exercised
live by the sample path. The save layer remains flag-only (detectRepetitionRisk);
this extends the engine's existing mutating collapse to the authoritative final boundary.

Verify: tsc clean, eslint clean, 57/57 affected unit tests, vite build OK.
